arch: drop -free-vectorize
Per #15 (paraphrased from Mingye Wang for commit msg):

Both the loop and SLP vectorizers are now enabled by -O2 in GCC 12. This
is equivalent with -ftree-vectorize. As a result, there is probably no
effect in specifying it explicitly.

This *might* also mean that we are accidentally vectorizing the Afterglow
ports, since instead of specifying `-fno-tree-XXX-vectorize', we only took
out `-ftree-vectorize' out.

P.S. GCC 12 also introduced a "very-cheap" cost model so it does not over-
vectorize at -O2.

Link: https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html#index-ftree-loop-vectorize
Link: https://www.phoronix.com/news/GCC-12-Auto-Vec-O2
Reported-by: Mingye Wang <arthur200126@gmail.com>