loongson3: Drop -mfix-loongson3-llsc
For Loongson 3 two workarounds for LL/SC may be needed:

    1:
    sync               # "LLSYNC"
    ll	$7,0($4)
    bne	$7,$3,2f
    li	$2,0
    move	$1,$6
    sc	$1,0($4)
    beq	$1,$0,1b
    li	$2,1
    2:
    sync               # "TGTSYNC"

The message of the AOSC patch adding -mfix-loongson3-llsc to GCC claims
"Binutils only handles LLSYNC, not TGTSYNC."  It's no longer correct:
the patch was written in 2018, but when the Binutils LL/SC errata
workaround was upstreamed in 2019 it actually handles both LLSYNC and
TGTSYNC.

As in abbs we already configure Binutils to enable the workaround by
default for loongson3, we don't need the GCC workaround anymore.  Drop
it here now, and it'll be dropped from our GCC patchset too.

Link: https://github.com/AOSC-Dev/aosc-os-abbs/blob/6e87aebe7f8d/app-devel/binutils/01-main/defines#L35
Link: https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=6f2117ba3892