arch/_common_switches.sh: append -O2 if AB_FLAGS_FTF is enabled
-D_FORTIFY_SOURCE requires optimisation (-O) and may cause build failures
if a build system passes only CPPFLAGS but not C/CXXFLAGS (i.e., Binutils
and GCC, and we are not sure if this has caused further silent failures
elsewhere).

Since it would not do much harm to pass our lowest optimisation level on
its own, pass -O2 as part of the AB_FLAGS_FTF condition.