arch: loongarch64_nosimd: Pass -mattr=-lsx to the linker plugin for Rust
The "Rust + linker-plugin = LSX in main" issue may haunt all Rust
binaries, even a "hello world" program is proven to be affected.  So
aosc-os-abbs/.../oma is just not the correct location to work it around.
And disabling LTO is overly aggressive, we just need to pass an option
to the linker plugin to disable LSX.

FWIW I'm unsure if we should report this as a Rustc issue.

Signed-off-by: Xi Ruoyao <xry111@xry111.site>